Move Minecraft Server base entity to its own file
